Asterisk-Unibo_vnfd/charms/native-charm/requirements.txt
ops >= 1.2.0
PyYAML
python-string-utils
